fix: rempved duplicated code
