+++ src/api/fetchersFunction/seoMetaDataFetchers.ts
@@ -1,63 +1,15 @@
 import axios from 'axios';
-import { CTP_API_URL, CTP_CUSTOM_OBJ_SEO_CONTAINER_KEY, CTP_CUSTOM_OBJ_SEO_CONTAINER_NAME, CTP_PROJECT_KEY, LS_KEY } from '../../constants';
+import {
+  CTP_API_URL,
+  CTP_CUSTOM_OBJ_SEO_CONTAINER_KEY,
+  CTP_CUSTOM_OBJ_SEO_CONTAINER_NAME,
+  CTP_PROJECT_KEY,
+  LS_KEY,
+} from '../../constants';
 import OpenAI from 'openai';
 import { getAllSavedRulesFromCtObj } from './ruleFetchers';
 import { getProductById } from './utils';
 
-export const generateSeoMetaData = async (
-  productId: string,
-  dataLocale: any,
-  setState?: Function
-) => {
-  const accessToken = localStorage.getItem(LS_KEY.CT_OBJ_TOKEN);
-  const openAiKey = localStorage.getItem(LS_KEY.OPEN_AI_KEY);
-  if (!openAiKey) {
-      setState?.((prev: any) => ({
-        ...prev,
-        notificationMessage:
-          'OpenAI key is missing. Please set it in the settings.',
-        notificationMessageType: 'error',
-      }));
-    return null;
-  }
-  try {
-    const productResponse = await getProductById(productId, dataLocale);
-
-    const productName = productResponse?.masterData?.current?.name;
-    const categories = productResponse?.masterData?.current?.categories;
-
-    const categoryNames = categories
-      ?.map((category: any) => category?.name)
-      ?.join(', ');
-    const query = `Product name: "${productName}", Categories: "${categoryNames}"`;
-
-    const localeQuery = dataLocale ? `, Locale: "${dataLocale}"` : '';
-    const data: any = await queryOpenAi(
-      query + localeQuery,
-      accessToken,
-      openAiKey
-    );
-    if (data?.status && data?.status == 401) {
-        setState?.((prev: any) => ({
-          ...prev,
-          notificationMessage: data?.error?.message,
-          notificationMessageType: 'error',
-        }));
-      return;
-    }
-
-    return { ...data, productId: productId };
-  } catch (error) {
-    console.error('Error generating SEO metadata:', error);
-
-      setState?.((prev: any) => ({
-        ...prev,
-        notificationMessage: 'Error generating SEO metadata.',
-        notificationMessageType: 'error',
-      }));
-    return null;
-  }
-};
 
 export const updateProductSeoMeta = async (
   productId: string,
@@ -118,18 +70,18 @@ export const updateProductSeoMeta = async (
   try {
     const response = await axios.post(apiUrl, payload, { headers });
 
-      setState?.((prev: any) => ({
-        ...prev,
-        notificationMessage: 'SEO title and description updated successfully.',
-        notificationMessageType: 'success',
-      }));
+    setState?.((prev: any) => ({
+      ...prev,
+      notificationMessage: 'SEO title and description updated successfully.',
+      notificationMessageType: 'success',
+    }));
     return response.data;
   } catch (error) {
-      setState?.((prev: any) => ({
-        ...prev,
-        notificationMessage: 'Error updating SEO title and description.',
-        notificationMessageType: 'error',
-      }));
+    setState?.((prev: any) => ({
+      ...prev,
+      notificationMessage: 'Error updating SEO title and description.',
+      notificationMessageType: 'error',
+    }));
     console.error('Error updating product SEO meta:', error);
     return null;
   }
@@ -146,8 +98,11 @@ export const queryOpenAi = async (
   });
   let updatedPrompt = '';
   if (accessToken) {
-    const prompt: any = getAllSavedRulesFromCtObj(accessToken, CTP_CUSTOM_OBJ_SEO_CONTAINER_NAME,
-      CTP_CUSTOM_OBJ_SEO_CONTAINER_KEY,);
+    const prompt: any = getAllSavedRulesFromCtObj(
+      accessToken,
+      CTP_CUSTOM_OBJ_SEO_CONTAINER_NAME,
+      CTP_CUSTOM_OBJ_SEO_CONTAINER_KEY
+    );
     const allEmpty = prompt?.value?.every((p: string) => /^\s*$/.test(p));
 
     // If any prompt is non-empty, update updatedPrompt
